+++ src/index.ts
@@ -1,7 +1,6 @@
 export { createIsolation } from './createIsolation';
 export {
-  createScope as INTERNAL_createScope,
-  Provider,
+  ScopeProvider,
   useAtom,
   useAtomValue,
   useSetAtom,